Only disable rx notifications on linux
